pre nego sto mi nestane struja

+++ apps/api/src/app/images/images.controller.ts
@@ -5,11 +5,13 @@ import {
   FileTypeValidator,
   FileValidator,
   Get,
+  InternalServerErrorException,
   MaxFileSizeValidator,
   Param,
   ParseFilePipe,
   Post,
   Res,
+  UnauthorizedException,
   UploadedFile,
   UseInterceptors,
 } from '@nestjs/common';
@@ -20,29 +22,59 @@ import { Response } from 'express';
 import { Public } from '../auth/decorators/public.decorator';
 import { Roles } from '../auth/decorators/roles.decorator';
 import { v4 as uuidv4 } from 'uuid';
+import { AccessTokenPayload, UserDto } from '@rwa/shared';
+import { ExtractUser } from '../auth/decorators/user.decorator';
+import { User } from '@rwa/entities';
+import { UserService } from '../user/user.service';
 
 @Controller('images')
 export class ImagesController {
-  constructor(private readonly imagesService: ImagesService) {}
+  constructor(
+    private readonly imagesService: ImagesService,
+    private userService: UserService
+  ) {}
 
-  @Roles(['admin'])
   @Post('upload')
   @UseInterceptors(
     FileInterceptor('file', {
       storage: diskStorage({
         destination: './uploads/',
         filename: (req, file, callback) => {
+          if (req.user == undefined) {
+            callback(new InternalServerErrorException(), '');
+            return;
+          }
+
+          const { user } = {
+            user: req.user,
+            iat: 0,
+            exp: 0,
+          } as AccessTokenPayload;
+
           const segments = file.originalname.split('.');
           if (segments.length < 2) {
             callback(new BadRequestException('Invalid image name'), '');
           } else {
-            callback(null, uuidv4().concat('.', segments[segments.length - 1]));
+            if (user.roles.includes('admin')) {
+              callback(
+                null,
+                uuidv4().concat('.', segments[segments.length - 1])
+              );
+            } else {
+              callback(null, `user_${user.id}`);
+            }
           }
         },
       }),
     })
   )
-  uploadFile(@UploadedFile() file: Express.Multer.File) {
+  async uploadImage(
+    @UploadedFile() file: Express.Multer.File,
+    @ExtractUser() user: User
+  ) {
+    if (!user.roles.includes('admin')) {
+      this.userService.setUserProfileImage(user.id, file.filename);
+    }
     return { name: file.filename };
   }
 
@@ -58,9 +90,16 @@ export class ImagesController {
     return this.imagesService.getNames();
   }
 
-  @Roles(['admin'])
   @Delete(':name')
-  remove(@Param('name') name: string) {
+  remove(@ExtractUser() user: User, @Param('name') name: string) {
+    if (!user.roles.includes('admin') && name != `user_${user.id}`) {
+      return new UnauthorizedException();
+    }
+
+    if (!user.roles.includes('admin')) {
+      this.userService.setUserProfileImage(user.id, null);
+    }
+
     return this.imagesService.remove(name);
   }
 }
